Read the chained sale back without a cast the engine has no word for
packages/provisioning's one failing suite was not waiting on the missing test harness, as the survey
had assumed: the three casts are in the TEST's own SQL, and the engine answers 'unrecognized token:
":"'. It reads three scalar subqueries — a count of the node's fiscal records, that sale's sequence
number, and the chain head's — and all three are integer values this engine hands back as numbers,
so the casts that turned PostgreSQL's bigint strings into numbers have nothing left to do.

The assertion is untouched. Control run before trusting the green: with the expected count changed
to 2 the case fails 'expected { registros: 1, … } to deeply equal { registros: 2, … }', so the
values really are the numbers 1 and not the strings "1" that would have made toEqual fail anyway.

packages/provisioning is now 13 files / 172 tests green, from 171 passed and 1 failed.

+++ packages/provisioning/src/venue-apply.e2e.test.ts
@@ -173,9 +173,9 @@ describe("a venue provisioned by applyVenue is immediately sellable", () => {
       head_secuencia: number;
     }>(sql`
       select
-        (select count(*) from registros_facturacion where node_id = ${venue.nodeId})::int as registros,
-        (select secuencia from registros_facturacion where sale_id = ${saleId})::int as secuencia,
-        (select secuencia from cadenas where node_id = ${venue.nodeId})::int as head_secuencia`);
+        (select count(*) from registros_facturacion where node_id = ${venue.nodeId}) as registros,
+        (select secuencia from registros_facturacion where sale_id = ${saleId}) as secuencia,
+        (select secuencia from cadenas where node_id = ${venue.nodeId}) as head_secuencia`);
     expect(chained.rows[0]).toEqual({ registros: 1, secuencia: 1, head_secuencia: 1 });
   });
 });
